+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
   ([!663](https://gitlab.com/fdroid/fdroidserver/merge_requests/663))
 * added support for gradle 5.5.1
   ([!656](https://gitlab.com/fdroid/fdroidserver/merge_requests/656))
+* add SHA256 to filename of repo graphics
+  ([!669](https://gitlab.com/fdroid/fdroidserver/merge_requests/669))
 
 ### Fixed
 * checkupdates: UpdateCheckIngore gets properly observed now